# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Cronuseo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/{proj_id}/resource_role/{id}": {
            "put": {
                "description": "Update resource role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resource Role"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "res_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Resource Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResourceRoleUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ResourceRole"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/{res_id}/resource_role": {
            "get": {
                "description": "Get all resource roles.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resource Role"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "res_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ResourceRole"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create resource role.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resource Role"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "res_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResourceRoleCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ResourceRole"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/{res_id}/resource_role/action/{resact_id}": {
            "post": {
                "description": "Assign resource role to group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resource Role"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "res_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Resource Action ID",
                        "name": "resact_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/{res_id}/resource_role/group/{group_id}": {
            "post": {
                "description": "Assign resource role to group.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resource Role"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "res_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/{res_id}/resource_role/user/{user_id}": {
            "post": {
                "description": "Assign resource role to user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resource Role"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "res_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/{res_id}/resource_role/{id}": {
            "get": {
                "description": "Get resource roles by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resource Role"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "res_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Resource Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResourceRoleWithGroupsUsers"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Delete resource role.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resource Role"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Resource ID",
                        "name": "res_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Resource Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.GroupOnlyWithID": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ResourceActionWithID": {
            "type": "object",
            "properties": {
                "resact_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ResourceRole": {
            "type": "object",
            "required": [
                "key",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string",
                    "minLength": 4
                },
                "name": {
                    "type": "string",
                    "minLength": 4
                }
            }
        },
        "models.ResourceRoleCreateRequest": {
            "type": "object",
            "required": [
                "key",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "key": {
                    "type": "string",
                    "minLength": 4
                },
                "name": {
                    "type": "string",
                    "minLength": 4
                }
            }
        },
...
        "models.ResourceRoleUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 4
                }
            }
        },
        "models.ResourceRoleWithGroupsUsers": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ResourceActionWithID"
                    }
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GroupOnlyWithID"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "res_id": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserOnlyWithID"
                    }
                }
            }
        },
...
        "models.UserOnlyWithID": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}